Product Composite api
- resttemplate을 사용해 서비스 간 api 호출
- api 정의 인터페이스 화

+++ api/src/main/java/microservices/api/core/product/Product.java
@@ -6,13 +6,13 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor
 public class Product {
-    private final Long productId;
+    private final int productId;
     private final String name;
     private final int weight;
     private final String serviceAddress;
 
     public Product() {
-        productId = 0L;
+        productId = 0;
         name = null;
         weight = 0;
         serviceAddress = null;
